# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MedVision AI",
    "description": "MedVision AI \u2014 an AI-powered healthcare education platform combining image, voice, and text understanding with a Retrieval-Augmented Generation medical knowledge base. AI output is informational only and is not a substitute for a licensed physician.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/admin/analytics": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Get Analytics",
        "operationId": "get_analytics_api_v1_admin_analytics_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetAnalyticsOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/admin/users": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "List Users",
        "operationId": "list_users_api_v1_admin_users_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "role",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/UserRole"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Role"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserOut"
                  },
                  "title": "Response List Users Api V1 Admin Users Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/users/{user_id}/role": {
      "put": {
        "tags": [
          "Admin"
        ],
        "summary": "Change User Role",
        "operationId": "change_user_role_api_v1_admin_users__user_id__role_put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          },
          {
            "name": "new_role",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/UserRole"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/users/{user_id}/deactivate": {
      "put": {
        "tags": [
          "Admin"
        ],
        "summary": "Deactivate User",
        "operationId": "deactivate_user_api_v1_admin_users__user_id__deactivate_put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/users/{user_id}/activate": {
      "put": {
        "tags": [
          "Admin"
        ],
        "summary": "Activate User",
        "operationId": "activate_user_api_v1_admin_users__user_id__activate_put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/doctors/{user_id}/verify": {
      "put": {
        "tags": [
          "Admin"
        ],
        "summary": "Verify Doctor",
        "operationId": "verify_doctor_api_v1_admin_doctors__user_id__verify_put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/knowledge-base/reindex": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Trigger Reindex",
        "operationId": "trigger_reindex_api_v1_admin_knowledge_base_reindex_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/admin/audit-logs": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Get Audit Logs",
        "operationId": "get_audit_logs_api_v1_admin_audit_logs_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response Get Audit Logs Api V1 Admin Audit Logs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/images/{image_id}/verify": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Verify Image Label",
        "description": "Expert validation step for the Image Dataset Builder / Active Learning pipeline.",
        "operationId": "verify_image_label_api_v1_admin_images__image_id__verify_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "image_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Image Id"
            }
          },
          {
            "name": "verified_label",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Verified Label"
            }
          },
          {
            "name": "severity",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Severity"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/models/retrain-vision": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Retrain Vision Model",
        "description": "Kicks off vision model retraining as a background Celery task.",
        "operationId": "retrain_vision_model_api_v1_admin_models_retrain_vision_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "DatasetAnalyticsOut": {
        "properties": {
          "total_diseases": {
            "type": "integer",
            "title": "Total Diseases"
          },
          "total_symptoms": {
            "type": "integer",
            "title": "Total Symptoms"
          },
          "total_medications": {
            "type": "integer",
            "title": "Total Medications"
          },
          "total_articles": {
            "type": "integer",
            "title": "Total Articles"
          },
          "total_images": {
            "type": "integer",
            "title": "Total Images"
          },
          "total_voice_records": {
            "type": "integer",
            "title": "Total Voice Records"
          },
          "total_lab_reports": {
            "type": "integer",
            "title": "Total Lab Reports"
          },
          "total_verified_cases": {
            "type": "integer",
            "title": "Total Verified Cases"
          },
          "total_users": {
            "type": "integer",
            "title": "Total Users"
          },
          "total_doctors": {
            "type": "integer",
            "title": "Total Doctors"
          },
          "total_patients": {
            "type": "integer",
            "title": "Total Patients"
          },
          "predictions_last_7_days": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Predictions Last 7 Days"
          },
          "uploads_last_7_days": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Uploads Last 7 Days"
          }
        },
        "type": "object",
        "required": [
          "total_diseases",
          "total_symptoms",
          "total_medications",
          "total_articles",
          "total_images",
          "total_voice_records",
          "total_lab_reports",
          "total_verified_cases",
          "total_users",
          "total_doctors",
          "total_patients",
          "predictions_last_7_days",
          "uploads_last_7_days"
        ],
        "title": "DatasetAnalyticsOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "UserOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "full_name": {
            "type": "string",
            "title": "Full Name"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "is_email_verified": {
            "type": "boolean",
            "title": "Is Email Verified"
          },
          "preferred_language": {
            "type": "string",
            "title": "Preferred Language"
          },
          "dark_mode": {
            "type": "boolean",
            "title": "Dark Mode"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "full_name",
          "role",
          "is_active",
          "is_email_verified",
          "preferred_language",
          "dark_mode",
          "created_at"
        ],
        "title": "UserOut"
      },
      "UserRole": {
        "type": "string",
        "enum": [
          "admin",
          "doctor",
          "patient"
        ],
        "title": "UserRole"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}